requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.16,<4.0
Pillow>=12.1,<13.0
asgiref>=3.8,<4.0
sqlparse>=0.5,<1.0
tzdata>=2025.3,<2027.0
celery>=5.4.0
redis>=4.5.0
openai>=1.3.0,<2.0
python-dotenv>=1.0.0
